Extract program-error crate (#2413)
* extract program-error crate

* remove `default-features = false` that doesn't do anything

* make serde optional in program-error crate

* update lock file

* remove thiserror from program-error crate

* update lock file

* move ProgramResult definition to program-error crate

* unused import

* fix features

* missing whitespace


---------
