+++ Cargo.toml
@@ -115,6 +115,7 @@ members = [
     "sdk/msg",
     "sdk/package-metadata-macro",
     "sdk/program",
+    "sdk/program-error",
     "sdk/program-memory",
     "sdk/program-option",
     "sdk/pubkey",
@@ -423,6 +424,7 @@ solana-perf = { path = "perf", version = "=2.1.0" }
 solana-poh = { path = "poh", version = "=2.1.0" }
 solana-poseidon = { path = "poseidon", version = "=2.1.0" }
 solana-program = { path = "sdk/program", version = "=2.1.0", default-features = false }
+solana-program-error = { path = "sdk/program-error", version = "=2.1.0" }
 solana-program-memory = { path = "sdk/program-memory", version = "=2.1.0" }
 solana-program-option = { path = "sdk/program-option", version = "=2.1.0" }
 solana-program-runtime = { path = "program-runtime", version = "=2.1.0" }
